+++ packages/core/src/utils/styles/index.ts
@@ -79,6 +79,7 @@ export * from '../breadcrumb-utils'
 export * from '../steps-utils'
 export * from '../pagination-utils'
 export * from '../dropdown-utils'
+export * from '../context-menu-utils'
 
 // Feedback components
 export * from '../drawer-utils'